backend/requirements.txt
fastapi>=0.68
uvicorn>=0.15
pydantic>=1.8
motor>=0.7
python-dotenv>=0.21
google-auth-oauthlib>=0.4
python-multipart>=0.0.5
pillow>=8.0
python-jose[cryptography]>=3.3.0